Fixed return type - CI

+++ clients/drcachesim/simulator/cache.cpp
@@ -62,7 +62,7 @@ cache_t::init(int associativity, int64_t line_size, int total_size,
     if (line_size < 1)
         return false;
     // convert total_size to num_blocks to fit for caching_device_t::init
-    int num_lines = total_size / line_size;
+    int num_lines = total_size / static_cast<int>(line_size);
 
     return caching_device_t::init(associativity, line_size, num_lines, parent, stats,
                                   prefetcher, inclusion_policy, coherent_cache, id,
